# contracts/nodes-v1.openapi.yaml
openapi: 3.1.0
info:
  title: RETINA node ingest
  version: 1.1.2
  description: |
    Tower Finder's HTTP API. The paths under `/v1/nodes` are the RETINA node ingest
    contract, generated from the server and versioned as a unit; everything else is
    internal to the map, the dashboard and the admin surfaces.

    ## Error taxonomy

    Every refusal under `/v1/nodes` carries `Error`, whose required key is `error`
    and whose optional `detail` names a field rather than describing it. The rest of
    this API answers in FastAPI's `{"detail": ...}`, which its own callers parse.

    | Status | Means |
    |---|---|
    | `400` | The body was refused. `invalid_config` for a configuration that failed validation, `invalid_body` for one that failed the schema |
    | `401` | The bearer token is bad, revoked or expired |
    | `403` | Registration refused, without saying why |
    | `409` | The frame names a `config_version` this server never issued |
    | `413` | The body exceeded the cap for its path |
    | `429` | Rate limited |

    `403` is deliberately opaque: unknown device, not yet accepted by Mender,
    already holding a valid token and in cooldown are one response, at one latency,
    with one `Retry-After`. Registration is limited per `node_id` and answers that
    same `403` rather than a `429`, since a distinct status would confirm to an
    unauthenticated caller that the identity it named is one the server tracks.

...

...

...
    someone's reading of a paragraph.

...

...

...

...

...
servers:
...
paths:
  /v1/nodes/register:
    post:
      tags:
      - registration
      summary: Register a node and obtain its bearer token.
      description: |-
        Unauthenticated. Called once, when the node has a `node_id` but no token.

        ## Preconditions

        The node must be enrolled with and accepted by Mender before it registers. The server asks
        Mender directly, while the request is open, so a node registering in the gap between
        enrolling and being accepted is refused and must retry.

        ## Failure handling

        Registration failures are deliberately opaque: unknown device, device not yet accepted by
        Mender, node already registered and holding a valid token, and an identity in cooldown all
        answer the same `403`, at the same latency, with the same `Retry-After`. Rate limiting
        answers it too. Each response carries `x-retry`, which is the field to act on.

        Registration is limited per `node_id` at 5 an hour and 20 a day. An ordinary node uses a
        handful of attempts in its lifetime, so this only bites a retry loop with no backoff, and
        the backoff is worth making real and jittered: a CGNAT pool rebooting together is a case
        the server is sized for, but a hot loop is not.

        ## Recovering a node that has lost its token

        A reflashed board comes back with the same `node_id`, a new Mender keypair and no token. On
        the wire that is indistinguishable from somebody enrolling a keypair of their own under that
        identity, so registering again revokes the previous token and is alerted on. There is
        nothing to build for it, but it is worth knowing when a reflashed board sits in a retry loop
        and the logs say nothing useful: the reason is on the server side.
      operationId: registerNode
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '200':
          description: Registered. Store the token, cache `node_ref`, and move on to heartbeat and detections.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: 'The configuration or the agreement records failed validation: `invalid_config`
            for a configuration value out of range, `invalid_body` for anything that failed the schema.
            `detail` names the offending field and nothing else, so retina-gui can put the message next
            to the input it belongs to. Retrying unchanged will not help.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          x-retry: never
          x-terminal: false
        '403':
          description: Refused, without saying why. Unknown device, not yet accepted by Mender, already
            holding a valid token, and an identity in cooldown are one response, at one latency, with
            one `Retry-After`. It is also the normal answer while Mender acceptance is pending, and the
            answer to a node registering too often, so a node that has just been flashed should expect
            it and keep retrying.
          headers:
            Retry-After:
              description: Seconds to wait before retrying. Honour it, then back off with jitter.
              required: true
              schema:
                type: integer
                maximum: 86400.0
                minimum: 0.0
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          x-retry: retry-after
          x-terminal: false
        '413':
          description: 'The body exceeded this path''s cap, which is checked from `Content-Length` before
            parsing. Per-field bounds in the schemas do not make this unreachable: `HeartbeatRequest.errors`
            alone permits 16 KiB of strings, `RegisterRequest.config` carries no bound at all, and JSON
            puts no length limit on a number''s text, so a schema-valid body can still exceed it.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          x-retry: never
          x-terminal: false
        5XX:
          description: 'Server side, though a 502 or 504 is the edge''s own response rather than the origin''s,
            and carries whatever body that gateway sends instead of this `Error`. Abandon the request
            and retry, backing off with jitter if it persists: on detection and heartbeat that retry is
            the node''s next frame or beat rather than a resend of this one, while registration and configuration
            retry the identical body.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          x-retry: backoff
          x-terminal: false
      x-cadence: once per node lifetime, plus operator reactivation
      x-max-body-bytes: 8192
...
components:
  schemas:
    AcceptanceRecord:
      properties:
        version:
          type: string
          maxLength: 32
          title: Version
        accepted_at:
          type: string
          format: date-time
          title: Accepted At
      additionalProperties: false
      type: object
      required:
      - version
      - accepted_at
      title: AcceptanceRecord
      description: One versioned thing the owner accepted, and when.
    Agreements:
      properties:
        licence:
          $ref: '#/components/schemas/AcceptanceRecord'
        remote_management:
          $ref: '#/components/schemas/AcceptanceRecord'
        publication:
          $ref: '#/components/schemas/PublicationChoice'
      additionalProperties: false
      type: object
      required:
      - licence
      - remote_management
      - publication
      title: Agreements
      description: Three separately versioned records, because they are withdrawn separately.
...
    ErrorBody:
      properties:
        error:
          type: string
          maxLength: 64
          title: Error
        detail:
          anyOf:
          - type: string
            maxLength: 512
          - type: 'null'
          title: Detail
      type: object
      required:
      - error
      title: ErrorBody
      description: |-
        The shape every node-API refusal wears. `error` is a stable slug;
        `detail`, present only when it applies, names the offending field or value
        rather than describing it. Registration's own refusals never carry
        `detail`, by design.
...
    PublicationChoice:
      properties:
        version:
          type: string
          maxLength: 32
          title: Version
        accepted_at:
          type: string
          format: date-time
          title: Accepted At
        choice:
          type: string
          enum:
          - public
          - private
          title: Choice
      additionalProperties: false
      type: object
      required:
      - version
      - accepted_at
      - choice
      title: PublicationChoice
      description: |-
        Whether the owner chose to publish this node's detections. `choice` is
        required, with no default: a node must send an explicit value.
    RegisterRequest:
      properties:
        node_id:
          type: string
          pattern: ^ret[0-9a-f]{8}$
          title: Node Id
        board_model:
          type: string
          maxLength: 64
          title: Board Model
        agreements:
          $ref: '#/components/schemas/Agreements'
        config:
          additionalProperties: true
          type: object
          title: Config
      additionalProperties: false
      type: object
      required:
      - node_id
      - board_model
      - agreements
      - config
      title: RegisterRequest
    RegisterResponse:
      properties:
        token:
          type: string
          maxLength: 128
          minLength: 32
          title: Token
        node_ref:
          type: string
          pattern: ^(nde|sim)[0-9a-z]{12}$
          title: Node Ref
        config_version:
          type: integer
          minimum: 1.0
          title: Config Version
        server_time:
          type: string
          format: date-time
          title: Server Time
      type: object
      required:
      - token
      - node_ref
      - config_version
      - server_time
      title: RegisterResponse
  securitySchemes:
    bearerAuth:
      type: http
      description: 'The token minted by `POST /v1/nodes/register`, persisted at mode 0600 under `/data`.
        Sent as `Authorization: Bearer <token>`. There is no expiry: a token dies by being revoked and
        by nothing else.'
      scheme: bearer